requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
redis==5.0.1
celery==5.3.6
pydantic[email]
kombu==5.3.5
amqp==5.1.1
PyQt5==5.15.9
requests==2.31.0
websocket-client==1.6.4
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1